daemon,pkg/option: Introduce config for socket traces
The config will be used to enable/disable socket
traces when socket-lb is enabled, independent of
packet related traces.
Per packet traces can produce significant number
of traces, and could lead to complexity issues,
so it's good to keep the two configs separate.

The socket tracing config is enabled by default
when socket-lb is enabled.

Signed-off-by: Aditi Ghag <aditi@cilium.io>