test: Little VM helper runner
This commit adds a runner for LVH VM images. The code is adapted from
the Tetragon runner and changed to be more generic. Code for the runner
has been split up into a package and the `test/lvh-runner` main file.

The package contains logic that modifies the given base image with a
set of actions like copying files and changing permissions. The package
generates a QEMU command to start the VM based on the given config.

The runner main file is very simple, converting flags into the config
used by the package. Thus the runner can be used to start a VM,
run scripts in there (using systemD to run them on startup for example).
But has no test-specific logic.

Having this separation allows for the implementation of more dedicated
testing tools to for example run matrix tests and process fail/pass
results.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>