bpf: Emit traces in bpf_sock
Use the socket tracing helpers introduced in previous commits
to send traces from the cgroup hooks that are executed on
connect and send/receive socket calls.

Traces will be sent in the forward as well as reverse path of
service translation performed as part of socket load-balancing.
Additionally, send them pre- and post- service translation in
order to report service clusterIP address and selected backend
IP address, respectively.
The cgroup id allows the agent to correlate these events to the
corresponding pod. While the socket cookie passed in the events
allows the agent to tie together multiple events associated with
a socket (e.g., forward and reverse service translation).

While Hubble needs to correlate socket tracing events at the pod level,
the BPF helper to get ancestor cgroup id of containers (which is
the cgroup id of the parent pod) is only available on kernel versions
>= 5.6. Hence, the datapath sends container level cgroup id in the
trace events, and the agent maps container metadata with the parent
pod metadata that Hubble can query.

Signed-off-by: Aditi Ghag <aditi@cilium.io>